Split common into server vs web-friendly versions
