+++ packages/common-web/README.md
@@ -1,0 +1,3 @@
+# ATP Common Library for Web
+
+A library containing code which is shared between web-friendly ATP packages.
