+++ packages/common-web/babel.config.js
@@ -1,0 +1,1 @@
+module.exports = require('../../babel.config.js')
