+++ packages/common-web/jest.config.js
@@ -1,0 +1,6 @@
+const base = require('../../jest.config.base.js')
+
+module.exports = {
+  ...base,
+  displayName: 'Common Web',
+}
