+++ packages/common-web/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@atproto/common",
+  "name": "@atproto/common-web",
   "version": "0.1.1",
-  "main": "dist/index.js",
+  "main": "src/index.ts",
   "license": "MIT",
   "scripts": {
     "test": "jest",
@@ -19,10 +19,8 @@
     "postpublish": "npm run update-main-to-src"
   },
   "dependencies": {
-    "@atproto/common-web": "*",
-    "@ipld/dag-cbor": "^7.0.3",
-    "cbor-x": "^1.5.1",
     "multiformats": "^9.6.4",
-    "pino": "^8.6.1"
+    "uint8arrays": "3.0.0",
+    "zod": "^3.14.2"
   }
 }
